Removing unused JSONString
